refactor (root): create src folder and move files inside
